requirements.txt
#Python 3.10
Flask==2.0.1
Werkzeug==2.0.3
ultralytics==8.3.108
opencv-python==4.11.0.86
folium==0.18.0
branca==0.8.1
torch==2.4.0
torchvision==0.19.0
torchaudio==2.2.1
timm==1.0.15
groundingdino==0.1.0
supervision==0.25.1
matplotlib==3.7.3
pandas==1.3.4
numpy==1.24.4
scipy==1.10.0
Pillow==10.4.0
requests==2.32.3
PyYAML==6.0.2
